pythomedic/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
opencv-python==4.9.0.80
numpy==1.26.4
paho-mqtt==1.6.1
firebase-admin==6.5.0
pydantic==2.7.1
